Refactoring and update of tiles dashboard
